+++ CHANGELOG.rst
@@ -36,6 +36,7 @@ Added
     - Display iRODS stats in details card (#188)
     - Ability to add extra flow parameters with ``get_extra_flow_data()`` (#297)
     - Script user workaround for non-working tickets in the proteomics use case (#297)
+    - Option for validating files without moving (#333)
 - **Projectroles**
     - Helper ``email.send_generic_mail()`` (#280)
     - Common template tag ``check_backend()``